test: extract child adapter fixtures and preserve readonly state
