+++ src/process/supervisor/adapters/child.test.ts
@@ -1,12 +1,16 @@
 // Child adapter tests cover adapting child processes to supervisor runs.
-import type { ChildProcess } from "node:child_process";
-import { EventEmitter } from "node:events";
 import fs from "node:fs";
-import { mkdir, writeFile } from "node:fs/promises";
+import { writeFile } from "node:fs/promises";
 import path from "node:path";
 import { PassThrough, Writable } from "node:stream";
 import { afterAll, afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 import { useAutoCleanupTempDirTracker } from "../../../../test/helpers/temp-dir.js";
+import {
+  createStubChild,
+  createWindowsNpmShim,
+  firstMockArg,
+  firstSpawnWithFallbackParams,
+} from "./child.test-support.js";
 import {
   expectRealExitWinsOverSigkillFallback,
   expectWaitStaysPendingUntilSigkillFallback,
@@ -52,51 +56,6 @@ let createChildAdapter: typeof import("./child.js").createChildAdapter;
 let getWindowsInstallRoots: typeof import("../../../infra/windows-install-roots.js").getWindowsInstallRoots;
 const tempDirs = useAutoCleanupTempDirTracker(afterEach);
 
-function createStubChild(pid = 1234) {
-  const child = new EventEmitter() as ChildProcess;
-  child.stdin = new PassThrough() as ChildProcess["stdin"];
-  child.stdout = new PassThrough() as ChildProcess["stdout"];
-  child.stderr = new PassThrough() as ChildProcess["stderr"];
-  Object.defineProperty(child, "stdio", {
-    value: [child.stdin, child.stdout, child.stderr],
-    configurable: true,
-  });
-  Object.defineProperty(child, "pid", { value: pid, configurable: true });
-  Object.defineProperty(child, "killed", { value: false, configurable: true, writable: true });
-  Object.defineProperty(child, "exitCode", { value: null, configurable: true, writable: true });
-  Object.defineProperty(child, "signalCode", { value: null, configurable: true, writable: true });
-  Object.defineProperty(child, "channel", { value: {}, configurable: true });
-  Object.defineProperty(child, "connected", { value: true, configurable: true, writable: true });
-  const killMock = vi.fn(() => true);
-  const sendMock = vi.fn((_message: unknown, ...args: unknown[]) => {
-    const callback = args.findLast((value) => typeof value === "function") as
-      | ((error: Error | null) => void)
-      | undefined;
-    callback?.(null);
-    return true;
-  });
-  const disconnectMock = vi.fn(() => {
-    Object.defineProperty(child, "connected", { value: false, configurable: true, writable: true });
-    child.emit("disconnect");
-  });
-  child.kill = killMock as ChildProcess["kill"];
-  child.send = sendMock as ChildProcess["send"];
-  child.disconnect = disconnectMock as ChildProcess["disconnect"];
-  const emitClose = (code: number | null, signal: NodeJS.Signals | null = null) => {
-    child.emit("close", code, signal);
-  };
-  const emitExit = (code: number | null, signal: NodeJS.Signals | null = null) => {
-    Object.defineProperty(child, "exitCode", { value: code, configurable: true, writable: true });
-    Object.defineProperty(child, "signalCode", {
-      value: signal,
-      configurable: true,
-      writable: true,
-    });
-    child.emit("exit", code, signal);
-  };
-  return { child, disconnectMock, killMock, sendMock, emitClose, emitExit };
-}
-
 async function createAdapterHarness(params?: {
   pid?: number;
   argv?: string[];
@@ -113,38 +72,6 @@ async function createAdapterHarness(params?: {
     stdinMode: "pipe-open",
   });
   return { adapter, killMock };
-}
-
-type SpawnWithFallbackParams = {
-  argv?: string[];
-  options?: {
-    detached?: boolean;
-    env?: NodeJS.ProcessEnv | Record<string, string>;
-    stdio?: string[];
-    windowsHide?: boolean;
-    windowsVerbatimArguments?: boolean;
-  };
-  fallbacks?: Array<{ options?: { detached?: boolean } }>;
-};
-
-function firstSpawnWithFallbackParams(): SpawnWithFallbackParams {
-  const [call] = spawnWithFallbackMock.mock.calls;
-  if (!call) {
-    throw new Error("expected spawnWithFallback call");
-  }
-  const [params] = call;
-  if (typeof params !== "object" || params === null || Array.isArray(params)) {
-    throw new Error("expected spawnWithFallback params to be an object");
-  }
-  return params;
-}
-
-function firstMockArg(mock: { mock: { calls: readonly unknown[][] } }, label: string): unknown {
-  const [call] = mock.mock.calls;
-  if (!call) {
-    throw new Error(`expected ${label} call`);
-  }
-  return call[0];
 }
 
 function expectedTrustedCmdExe(): string {
@@ -209,29 +136,10 @@ describe("createChildAdapter", () => {
     vi.useRealTimers();
   });
 
-  const createWindowsNpmShim = async (params: { command: string; packagePath: string[] }) => {
-    const binDir = tempDirs.make("openclaw-child-shim-");
-    const entrypoint = path.join(binDir, "node_modules", ...params.packagePath);
-    await mkdir(path.dirname(entrypoint), { recursive: true });
-    await writeFile(entrypoint, "", "utf8");
-    const relativeEntrypoint = path.relative(binDir, entrypoint).replaceAll(path.sep, "\\");
-    const shimHead =
-      "@ECHO off\r\nGOTO start\r\n:find_dp0\r\nSET dp0=%~dp0\r\nEXIT /b\r\n:start\r\nSETLOCAL\r\nCALL :find_dp0\r\n";
-    const shimCommand = entrypoint.endsWith(".exe")
-      ? `"%dp0%\\${relativeEntrypoint}" %*\r\n`
-      : `IF EXIST "%dp0%\\node.exe" (\r\n  SET "_prog=%dp0%\\node.exe"\r\n) ELSE (\r\n  SET "_prog=node"\r\n)\r\nendLocal & goto #_undefined_# 2>NUL || title %COMSPEC% & "%_prog%" "%dp0%\\${relativeEntrypoint}" %*\r\n`;
-    await writeFile(
-      path.join(binDir, `${params.command}.cmd`),
-      `${shimHead}${shimCommand}`,
-      "utf8",
-    );
-    return { binDir, entrypoint };
-  };
-
   it("uses process-tree kill for default SIGKILL", async () => {
     const { adapter, killMock } = await createAdapterHarness({ pid: 4321 });
 
-    const spawnArgs = firstSpawnWithFallbackParams();
+    const spawnArgs = firstSpawnWithFallbackParams(spawnWithFallbackMock);
     // On Windows, detached defaults to false (headless Scheduled Task compat);
     // on POSIX, detached is true with a no-detach fallback.
     if (process.platform === "win32") {
@@ -267,9 +175,10 @@ describe("createChildAdapter", () => {
       input: "{}",
     });
 
-    expect(firstSpawnWithFallbackParams().options?.detached).toBe(process.platform !== "win32");
-    expect(firstSpawnWithFallbackParams().fallbacks).toEqual([]);
-    expect(firstSpawnWithFallbackParams().options?.stdio).toEqual(["pipe", "pipe", "pipe", "ipc"]);
+    const spawnArgs = firstSpawnWithFallbackParams(spawnWithFallbackMock);
+    expect(spawnArgs.options?.detached).toBe(process.platform !== "win32");
+    expect(spawnArgs.fallbacks).toEqual([]);
+    expect(spawnArgs.options?.stdio).toEqual(["pipe", "pipe", "pipe", "ipc"]);
 
     await adapter.openStartGate?.();
     expect(sendMock).toHaveBeenCalledWith(
@@ -292,7 +201,11 @@ describe("createChildAdapter", () => {
       const { child, disconnectMock, emitExit, killMock } = createStubChild();
       // Node marks connected false before a queued IPC disconnect has completed.
       disconnectMock.mockImplementation(() => {
-        child.connected = false;
+        Object.defineProperty(child, "connected", {
+          value: false,
+          configurable: true,
+          writable: true,
+        });
       });
       spawnWithFallbackMock.mockResolvedValue({ child, usedFallback: false });
       const adapter = await createChildAdapter({
@@ -556,7 +469,7 @@ describe("createChildAdapter", () => {
       },
     });
 
-    expect(firstSpawnWithFallbackParams().options?.stdio).toEqual([
+    expect(firstSpawnWithFallbackParams(spawnWithFallbackMock).options?.stdio).toEqual([
       "pipe",
       "pipe",
       "pipe",
@@ -615,7 +528,9 @@ describe("createChildAdapter", () => {
       },
     });
 
-    expect(firstSpawnWithFallbackParams().options?.stdio?.[3]).toBe("overlapped");
+    expect(firstSpawnWithFallbackParams(spawnWithFallbackMock).options?.stdio?.[3]).toBe(
+      "overlapped",
+    );
   });
 
   it("passes detached:false to signalProcessTree when spawn fell back to no-detach (#71662 follow-up)", async () => {
@@ -719,7 +634,7 @@ describe("createChildAdapter", () => {
       argv: ["node", "-e", "setTimeout(() => {}, 1000)"],
     });
 
-    const spawnArgs = firstSpawnWithFallbackParams();
+    const spawnArgs = firstSpawnWithFallbackParams(spawnWithFallbackMock);
     expect(spawnArgs.options?.stdio?.[0]).toBe("inherit");
     expect(adapter.stdin).toBeUndefined();
   });
@@ -972,7 +887,7 @@ describe("createChildAdapter", () => {
 
     await createAdapterHarness({ pid: 7777 });
 
-    const spawnArgs = firstSpawnWithFallbackParams();
+    const spawnArgs = firstSpawnWithFallbackParams(spawnWithFallbackMock);
     expect(spawnArgs.options?.detached).toBe(false);
     expect(spawnArgs.fallbacks ?? []).toStrictEqual([]);
     expect(createServiceChildRelayAdapterMock).not.toHaveBeenCalled();
@@ -986,7 +901,7 @@ describe("createChildAdapter", () => {
       argv: ["node", "-e", "process.exit(0)"],
     });
 
-    const spawnArgs = firstSpawnWithFallbackParams();
+    const spawnArgs = firstSpawnWithFallbackParams(spawnWithFallbackMock);
     expect(spawnArgs.argv).toEqual(["node", "-e", "process.exit(0)"]);
     expect(spawnArgs.options?.env).toBeUndefined();
   });
@@ -1000,7 +915,7 @@ describe("createChildAdapter", () => {
       env: { PATH: "", PATHEXT: ".EXE;.CMD;.BAT" },
     });
 
-    const spawnArgs = firstSpawnWithFallbackParams();
+    const spawnArgs = firstSpawnWithFallbackParams(spawnWithFallbackMock);
     expect(spawnArgs.argv).toEqual([
       expectedTrustedCmdExe(),
       "/d",
@@ -1017,6 +932,7 @@ describe("createChildAdapter", () => {
   it("unwraps Gemini's npm shim and preserves prompt argv on Windows", async () => {
     setPlatform("win32");
     const { binDir, entrypoint } = await createWindowsNpmShim({
+      binDir: tempDirs.make("openclaw-child-shim-"),
       command: "gemini",
       packagePath: ["@google", "gemini-cli", "bundle", "gemini.js"],
     });
@@ -1030,7 +946,7 @@ describe("createChildAdapter", () => {
       env: { PATH: binDir, PATHEXT: ".EXE;.CMD;.BAT" },
     });
 
-    const spawnArgs = firstSpawnWithFallbackParams();
+    const spawnArgs = firstSpawnWithFallbackParams(spawnWithFallbackMock);
     expect(spawnArgs.argv?.[0]?.toLowerCase()).toBe(nodePath.toLowerCase());
     expect(spawnArgs.argv?.slice(1)).toEqual([entrypoint, "--prompt", prompt]);
     expect(spawnArgs.options?.windowsVerbatimArguments).toBeUndefined();
@@ -1040,6 +956,7 @@ describe("createChildAdapter", () => {
   it("unwraps Claude's npm shim to its native executable on Windows", async () => {
     setPlatform("win32");
     const { binDir, entrypoint } = await createWindowsNpmShim({
+      binDir: tempDirs.make("openclaw-child-shim-"),
       command: "claude",
       packagePath: ["@anthropic-ai", "claude-code", "bin", "claude.exe"],
     });
@@ -1050,7 +967,7 @@ describe("createChildAdapter", () => {
       env: { PATH: binDir, PATHEXT: ".EXE;.CMD;.BAT" },
     });
 
-    const spawnArgs = firstSpawnWithFallbackParams();
+    const spawnArgs = firstSpawnWithFallbackParams(spawnWithFallbackMock);
     expect(spawnArgs.argv).toEqual([entrypoint, "--version"]);
     expect(spawnArgs.options?.windowsVerbatimArguments).toBeUndefined();
     expect(spawnArgs.fallbacks).toStrictEqual([]);
@@ -1090,7 +1007,7 @@ describe("createChildAdapter", () => {
       }
     }
 
-    const spawnArgs = firstSpawnWithFallbackParams();
+    const spawnArgs = firstSpawnWithFallbackParams(spawnWithFallbackMock);
     expect(spawnArgs.argv?.slice(0, 4)).toEqual([
       "/bin/sh",
       "-c",
@@ -1123,7 +1040,7 @@ describe("createChildAdapter", () => {
       restoreLinuxShell();
     }
 
-    const spawnArgs = firstSpawnWithFallbackParams();
+    const spawnArgs = firstSpawnWithFallbackParams(spawnWithFallbackMock);
     expect(spawnArgs.argv).toEqual(["/usr/bin/node", "-e", "process.exit(0)"]);
     expect(spawnArgs.options?.env).toEqual({ HOME: "/worker-home", PATH: "/usr/bin" });
   });
@@ -1135,7 +1052,7 @@ describe("createChildAdapter", () => {
       env: { FOO: "bar", COUNT: "12", DROP_ME: undefined },
     });
 
-    const spawnArgs = firstSpawnWithFallbackParams();
+    const spawnArgs = firstSpawnWithFallbackParams(spawnWithFallbackMock);
     expect(spawnArgs.options?.env).toEqual({ FOO: "bar", COUNT: "12" });
   });
